fix fib test

+++ src/test/app/wasm_fixtures/fib.c
@@ -1,5 +1,5 @@
-typedef long long mint;
-// typedef int mint;
+// typedef long long mint;
+typedef int mint;
 
 mint
 fib(mint n)
